Add Synchronizer example
