+++ crates/rmf_site_sim/src/compute.rs
@@ -1,7 +1,9 @@
 use crate::event::CandidateDiscreteEvents;
 use crate::schedule::{SimulationStartup, SimulationSystemExec};
-use crate::simulation::{SimulationComputeState, SimulationComputeUpdate, SimulationPluginFactory};
-use crate::sync::{SimulationEventBuffer, SimulationState, StateUpdateSender, Synchronizer};
+use crate::simulation::{
+    SimulationComputeState, SimulationComputeUpdate, SimulationPluginFactory, SimulationState,
+};
+use crate::sync::{SimulationEventBuffer, StateUpdateSender, Synchronizer};
 use crate::time::SimulationTime;
 use bevy::prelude::*;
 use bevy::tasks::{AsyncComputeTaskPool, TaskPool};
